Courss ended, is this the end?

+++ ceii-feed/controllers/api/Post.js
@@ -1,4 +1,5 @@
 const PostService = require("../../services/Post");
+const UserService = require("../../services/User");
 const {verifyID} = require ("../../utils/MongoUtils")
 const {verifyTypeNumber} = require("../../utils/MiscUtils");
 const controller={};
@@ -13,7 +14,8 @@ controller.create= async (req, res)=>{
     }
 
     try{
-        const createPost = await PostService.create(body);
+        const { user } = req;
+        const createPost = await PostService.create(req.body,user._id);
         if(!createPost.success){
             return res.status(409).json(createPost.content);
         }
@@ -41,8 +43,35 @@ controller.finOneById = async (req, res) =>{
     }
 };
 
+controller.findAllByUser = async (req, res) => {
+    const {id = req.user._id} = req.query;
+
+    if(!verifyID(id)){
+        return res.status(400).json({
+            error: "Error in ID"
+        });
+    }
+
+    try{
+        const userExists = await UserService.findOneById(id);
+        if(!userExists){
+            return res.status(404).json(userExists.content);
+        }
+
+        const postsByUser = await PostService.findAllByUserID(id);
+        return res.status(200).json(postsByUser.content);
+
+    }catch (e) {
+        return res.status(500).json({
+            error:"Internal Server Error"
+        })
+    }
+}
+
 controller.findAll = async (req,res) => {
     const {page =0, limit=10} = req.query;
+
+    debug("Hola soy un usuario" + req.user);
 
     if(!verifyTypeNumber(page,limit)){
         return res.status(400).json({
@@ -104,6 +133,14 @@ controller.updatePost = async (req,res) => {
         if(!postExists.success){
             return res.status(404).json(postExists.content);
         }
+
+        const {user} = req;
+        const myPost = PostService.verifyUserAuthority(postExists.content, user);
+
+        if(!myPost.success){
+            return res.status(401).json(myPost.content);
+        }
+
         const postUpdated = await PostService.updateOneById(
             postExists.content,
             fieldVerified.content
@@ -138,6 +175,14 @@ controller.deleteOneByID = async (req,res) =>{
         if(!postExists.success){
             return res.status(404).json(postExists.content);
         }
+
+        const {user} = req;
+        const myPost = PostService.verifyUserAuthority(postExists.content, user);
+
+        if(!myPost.success){
+            return res.status(401).json(myPost.content);
+        }
+
         const deleted = await PostService.deleteOneById(_id);
         if(!deleted.success){
             return res.status(409).json(deleted.content);
@@ -149,6 +194,6 @@ controller.deleteOneByID = async (req,res) =>{
         })
     }
 
-}
+};
 
 module.exports = controller;
